fix(Exercise): Refactor
